+++ tests/unit_tests/model/test_model_events.py
@@ -347,7 +347,6 @@ class TestEvents:
             last_qty=Quantity.from_str("0.561000"),
             last_px=Price.from_str("15600.12445"),
             currency=USDT,
-            is_inverse=False,
             commission=Money("12.20000000", USDT),
             liquidity_side=LiquiditySide.MAKER,
             execution_ns=0,